refactor(android): merge PushInstanceIDListenerService into FCMService
Since 1ae3428103298b85263785f1055b04b277feb92a, when
`PushInstanceIDListenerService` was changed away from the deprecated
`FirebaseInstanceIdService` to use
`FirebaseMessagingService::onNewToken`, there is no longer a need to use
a separate service anymore. Further, the `INSTANCE_ID_EVENT` intent
filter is not needed.
