+++ src/android/com/adobe/phonegap/push/FCMService.kt
@@ -22,6 +22,7 @@ import com.adobe.phonegap.push.PushPlugin.Companion.isActive
 import com.adobe.phonegap.push.PushPlugin.Companion.isInForeground
 import com.adobe.phonegap.push.PushPlugin.Companion.sendExtras
 import com.adobe.phonegap.push.PushPlugin.Companion.setApplicationIconBadgeNumber
+import com.google.firebase.iid.FirebaseInstanceId
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
 import org.json.JSONArray
@@ -61,6 +62,24 @@ class FCMService : FirebaseMessagingService() {
       PushConstants.COM_ADOBE_PHONEGAP_PUSH,
       MODE_PRIVATE
     )
+
+  /**
+   * Called when a new token is generated, after app install or token changes.
+   *
+   * @param token
+   */
+  override fun onNewToken(token: String) {
+    super.onNewToken(token)
+    FirebaseInstanceId.getInstance().instanceId
+      .addOnSuccessListener { instanceIdResult ->
+        // Get updated InstanceID token.
+        val refreshedToken = instanceIdResult.token
+        Log.d(TAG, "Refreshed token: $refreshedToken")
+
+        // TODO: Implement this method to send any registration to your app's servers.
+        //sendRegistrationToServer(refreshedToken);
+      }
+  }
 
   /**
    * Set Notification
